It has come to my attention that I had accidentally pushed the wrong tree here that was based on an older version. 

This is the correct 5.08 tree.

Updated the resources submodule as well.

Including the new installer as well with the skeleton directory tree that is used in the VS solution to stage precompiled libs/plugins etc to the output dir (the final package directory tree basically) that the installer uses as source. This may not be of use to a lot of people (as well as the new inno installer) but trying to avoid pushing these to the repo is how the wrong tree was pushed to begin with. It easier for me not to have to deal with separating things in order to keep this updated. I will merge the commits from stock after 5.08, once time allows again.

+++ scripting/methods/methods_arrays.cpp
@@ -175,7 +175,7 @@ VARIANT CMUSHclientDoc::ArrayListAll()
   // put the arrays into the array
   if (!GetArrayMap ().empty ())    // cannot create empty dimension
     {
-    sa.CreateOneDim (VT_VARIANT, GetArrayMap ().size ());
+    sa.CreateOneDim (VT_VARIANT, (DWORD) GetArrayMap ().size ());
 
     for (tStringMapOfMaps::iterator it = GetArrayMap ().begin (); 
          it != GetArrayMap ().end ();
@@ -208,7 +208,7 @@ VARIANT CMUSHclientDoc::ArrayListKeys(LPCTSTR Name)
   // put the keys into the array
   if (!it->second->empty ())    // cannot create empty dimension
     {
-    sa.CreateOneDim (VT_VARIANT, it->second->size ());
+    sa.CreateOneDim (VT_VARIANT, (DWORD) it->second->size ());
 
     for (tStringToStringMap::iterator i = it->second->begin (); 
          i != it->second->end ();
@@ -241,7 +241,7 @@ VARIANT CMUSHclientDoc::ArrayListValues(LPCTSTR Name)
   // put the keys into the array
   if (!it->second->empty ())    // cannot create empty dimension
     {
-    sa.CreateOneDim (VT_VARIANT, it->second->size ());
+    sa.CreateOneDim (VT_VARIANT, (DWORD) it->second->size ());
 
     for (tStringToStringMap::iterator i = it->second->begin (); 
          i != it->second->end ();
@@ -562,7 +562,7 @@ BOOL CMUSHclientDoc::ArrayKeyExists(LPCTSTR Name, LPCTSTR Key)
 // return count of all arrays
 long CMUSHclientDoc::ArrayCount() 
 {
-	return GetArrayMap ().size ();
+	return (long) GetArrayMap ().size ();
 } // end of CMUSHclientDoc::ArrayCount
 
 // return count of elements in a particular array
@@ -574,7 +574,7 @@ long CMUSHclientDoc::ArraySize(LPCTSTR Name)
   if (it == GetArrayMap ().end ())
     return 0;   
 
-	return it->second->size ();
+	return (long) it->second->size ();
 }   // end of CMUSHclientDoc::ArraySize
 
 VARIANT CMUSHclientDoc::ArrayGetFirstKey(LPCTSTR Name)
